requirements.txt
kafka-python==1.4.6
psycopg2==2.8.3
SQLAlchemy==1.3.7
PyYAML==5.1.2
loguru==0.3.2
